tweak(client): check that we're initialized before trying proximity
This makes sure we're connected and fully initialized (we have a channel) before
we try to do proximity adds, since we won't be able to hear them while
we're initializing

Add checks for our usage of MumbleGetVoiceChannelFromServerId so we
don't try to add to them as a target if they're not currently in mumble.

Add retry logic for when in spectator mode and a player fails to get
added via `onPlayerJoined` so we'll eventually listen to them if they
connect to mumble.

This is untested, this is meant to fix #524